requirements.txt
# airflow packages
apache-airflow-providers-postgres

# standard Python packages
pandas
numpy